fix:파일 합치기, 확률 넣기
